ml-server/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pillow>=10.0.0
numpy>=1.24.0
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
sentence-transformers>=2.2.0
opencv-python>=4.8.0
scikit-learn>=1.3.0
pydantic>=2.5.0
python-dotenv>=1.0.0
requests>=2.31.0
aiofiles>=23.2.0

